src/main/java/kitchenpos/domain/NumberOfGuests.java
package kitchenpos.domain;

import kitchenpos.common.ErrorCode;

import javax.persistence.Column;
import javax.persistence.Embeddable;
import java.util.Objects;

@Embeddable
public class NumberOfGuests {

    private static final int ZERO = 0;

    @Column(nullable = false, columnDefinition = "INT(11)")
    private int numberOfGuests;

    protected NumberOfGuests() {
    }

    private NumberOfGuests(int numberOfGuests) {
        isValidNumberOfGuestsIsNegative(numberOfGuests);
        this.numberOfGuests = numberOfGuests;
    }

    public static NumberOfGuests from(int numberOfGuests) {
        return new NumberOfGuests(numberOfGuests);
    }

    private void isValidNumberOfGuestsIsNegative(int numberOfGuests) {
        if (numberOfGuests < ZERO) {
            throw new IllegalArgumentException(ErrorCode.NUMBER_OF_GUEST_NOT_NEGATIVE.getMessage());
        }
    }

    public int getNumberOfGuests() {
        return numberOfGuests;
    }

    @Override
    public boolean equals(Object o) {
        if (this == o) return true;
        if (o == null || getClass() != o.getClass()) return false;
        NumberOfGuests that = (NumberOfGuests) o;
        return numberOfGuests == that.numberOfGuests;
    }

    @Override
    public int hashCode() {
        return Objects.hash(numberOfGuests);
    }

}
